requirement.txt
flask
flask-cors
python-dotenv
flask-socketio
boto3
flask_jwt_extended
requests
yarl